Context casi listo falta La pagina de login ya register funciona con firebase
