--- src/components/Admin/DashBoard.js
@@ -1,8 +1,0 @@
-import React from 'react'
-const DashBoard = () => {
-  return (
-    <div><h1>DashBoard</h1></div>
-  )
-}
-
-export default DashBoard
