+++ src/components/Login/Login.js
@@ -1,41 +1,19 @@
-import React from "react"
+import React from "react";
 import {makeStyles} from "@material-ui/core"
-import {useState} from "react"
-import {createUserWithEmailAndPassword, signInWithEmailAndPassword} from "firebase/auth"
-import {doc, setDoc} from "firebase/firestore"
-import {auth, db, googleProvider, facebookProvider, twitterProvider, githubProvider} from "../../firebase/credenciales"
+import { useNavigate } from 'react-router-dom';
+import { sessionContext } from '../../context/SessionContext';
+//Falta manejar esto por context
 
 const Login = () => {
+
+  const navigate = useNavigate()
   const classes=useStyle()
-  const [isRegistrando,setIsRegistrando]=useState(false)
-  
+  // const {session, login} = sessionContext()
 
-  async function registrarUsuario(email,password){
-    const infoUsuario= await createUserWithEmailAndPassword(
-      auth,
-      email,
-      password
-      ).then((usuarioFirebase)=>{
-      return usuarioFirebase;
-    });
-    console.log(infoUsuario)
-
-
-    const docuRef =  doc(db,`usuarios/${infoUsuario.user.uid}`)
-    setDoc(docuRef, {correo:email,rol:"usuario"})
-
-  } 
-  function submitHandler(e){
-    e.preventDefault();
-    const email = e.target.elements.email.value
-    const password = e.target.elements.password.value;
-    console.log(email,password)
-    if(isRegistrando){
-      registrarUsuario(email,password);
-    }else{
-      signInWithEmailAndPassword(auth,email,password)
-    }
-
+  const submitHandler = (e) => {
+    e.preventDefault()
+    const [email, password] = e.target
+    navigate("/")
   }
 
   // const handleGoogleLogin = async () => {
@@ -55,7 +33,6 @@ const Login = () => {
   //   sessionCheck(response)
   // };
   
-  
   return (
     <div>
       <form onSubmit={submitHandler}>
@@ -68,7 +45,7 @@ const Login = () => {
           <input type="password" id='password'/>
           <input
             type="submit"
-            value={isRegistrando ? "Registrar" : "Iniciar Sesión"}
+            value={ "Iniciar Sesión"}
           />
         </label>
         {/* <button className={styles.submits} type="button" onClick={handleGoogleLogin}>
@@ -84,19 +61,13 @@ const Login = () => {
           <img className={styles.iconPic} src =" "></img>
         </button>      */}
       </form>
-      <button onClick={()=> setIsRegistrando(!isRegistrando)}>
-        {isRegistrando ? "Ya Tengo Una Cuenta" : "Quiero Registrarme"}
-      </button>
-
-
+      <button onClick={()=> navigate("/register")}>Registrate aqui</button>
     </div>
-    
   )
-
 }
 
 const useStyle=makeStyles((theme)=>({
 }))
 
 
-export default Login
+export default Login;
