+++ src/pages/DashboardCiudadesPage.js
@@ -1,3 +1,3 @@
 // No estoy seguro de esta Pagina porque quizas podamos poner los dashboards 
 // en las mismas paginas de hotel y ciudades
-
+import React from "react";
