+++ src/pages/DashboardHotelesPage.js
@@ -1,0 +1,4 @@
+// No estoy seguro de esta Pagina porque quizas podamos poner los dashboards 
+// en las mismas paginas de hotel y ciudades
+import React from "react";
+
