+++ src/pages/HabListaPage.js
@@ -2,3 +2,4 @@
 //aqui es donde iria el Date Picker con su respectiva habitacion la pregunta 
 //es si cada habitacion al darle reservar
 //Eliges las fechas y la reservas y pagas en la misma pagina???????
+import React from "react";
