+++ src/pages/RegisterPage.js
@@ -1,5 +1,5 @@
 //Aqui llegarian desde el login???
-
+import React from "react";
 import Register from "../components/Register/Register"
